kernel/spec-nic.h: fix macro for 64-bit build
SET_HI32(variable, highhalf) was meant to prevent warnings and
unneeded code to get the best of 32 and 64 bits. But I had a bug
in 64-bit mode. This is the trivial fix.

+++ kernel/spec-nic.h
@@ -62,7 +62,7 @@
 /* This is somehow generic, but I find no better place at this time */
 #ifndef SET_HI32
 #  if BITS_PER_LONG > 32
-#    define SET_HI32(var, value)	((var) |= (value) << 32)
+#    define SET_HI32(var, value)	((var) |= (long)(value) << 32)
 #    define GET_HI32(var)		((var) >> 32)
 #  else
 #    define SET_HI32(var, value)	((var) |= 0)
